refactor(util): util 패키지 삭제 후 각 파일을 더 구체적인 이름의 패키지로 옮김
